+++ git-repository/src/easy/ext/mod.rs
@@ -6,3 +6,6 @@ pub use reference::ReferenceAccessExt;
 
 mod repository;
 pub use repository::RepositoryAccessExt;
+
+mod cache;
+pub use cache::CacheAccessExt;
