+++ git-repository/tests/easy/ext/object.rs
@@ -16,56 +16,44 @@ mod write_object {
 }
 
 mod find {
-    use git_repository::prelude::{ReferenceAccessExt, RepositoryAccessExt};
+    use git_repository::prelude::{CacheAccessExt, ReferenceAccessExt};
 
     #[test]
-    fn find_and_try_find_with_and_without_object_cache() {
-        let repo = crate::basic_repo().unwrap();
+    fn find_and_try_find_with_and_without_object_cache() -> crate::Result {
+        let repo = crate::basic_repo()?;
         for round in 1..=2 {
             match round {
                 1 => assert!(
-                    repo.object_cache_size(None).unwrap().is_none(),
+                    repo.object_cache_size(None)?.is_none(),
                     "default is to have no object cache"
                 ),
                 2 => {
-                    repo.object_cache_size(128 * 1024).unwrap();
+                    repo.object_cache_size(128 * 1024)?;
                 }
                 _ => unreachable!("BUG"),
             }
-            for commit_id in repo
-                .head()
-                .unwrap()
-                .peeled()
-                .unwrap()
-                .id()
-                .expect("born")
-                .ancestors()
-                .unwrap()
-                .all()
-            {
-                let commit = commit_id.unwrap();
-                assert_eq!(commit.object().unwrap().kind, git_object::Kind::Commit);
+            for commit_id in repo.head()?.peeled()?.id().expect("born").ancestors()?.all() {
+                let commit = commit_id?;
+                assert_eq!(commit.object()?.kind, git_object::Kind::Commit);
                 if round == 2 {
                     assert_eq!(
-                        commit.object().unwrap().kind,
+                        commit.object()?.kind,
                         git_object::Kind::Commit,
                         "repeated request triggers cache and doesn't fail"
                     );
                 }
-                assert_eq!(
-                    commit.try_object().unwrap().expect("exists").kind,
-                    git_object::Kind::Commit,
-                );
+                assert_eq!(commit.try_object()?.expect("exists").kind, git_object::Kind::Commit,);
             }
 
             if round == 2 {
                 assert_eq!(
-                    repo.object_cache_size(None).unwrap(),
+                    repo.object_cache_size(None)?,
                     Some(128 * 1024),
                     "it returns the previous cache"
                 );
             }
         }
+        Ok(())
     }
 }
